Crusader Perks Check-In
From the Crusader Perks Page, there is now a page to see how many events you have checked in to. This was able to happen because of a function in app.js called countCEvents that counts the amount of events the user has checked into. There is also a Check In Alert Popup in event.js now so the add event popup isn't used for checking in anymore. Additionally, a typo was fixed in the crusader perks html